Do work at low res, capture at high
Working specifically to the 1920x1080 camera we have for the project, we
now scale the face detection area and grab the image at a higher res.

Also added back some capabilities that were lost in a refactor to base
the thresholding only on an area centered about the face.

I also don't think you desparately need a picture of my mug.